Optimise formatting of metrics for librato
Previously this was making sure that all tags were recorded in ascending
order. This wasn't stricyl necessary - we just need to ensure that
applying the formatter multiple times to the same map of tags produces
consistent output so that our tests pass consistently.

We can do this by just sorting the generated tag pairs.