Add initial support for extracting tags from envoy metrics
The previous implementation of extracting tags would replace the entire
name of the metric every time a tag was extracted. This works fine for
consul metrics where we're operating on very specific metrics that don't
share tags.

Envoy uses the same convention for tags across many different metrics.
We want to be able to define one rule and have it extract the tags
from all metrics. To do this we needed to change the rule matcher to
only modify the part of the metric name that was matched, rather than
the whole thing.