Add support for extracting tags from a specific part of a metric name
Some metrics in envoy contain tag-like information such as the http
response status code (200), or response status class (2XX).

The previous implementation assumed tag values would be delimited by
periods (e.g. `cluster.{cluster_name}.foo.bar`), but in this case we
have things like `_rq_2XX` and `_rq_200`.

This commit changes the algorithm to support that. It also removes
support for the wildcard `*` operator. I had copied it from [seatgeek's
implementation](https://github.com/seatgeek/statsd-rewrite-proxy), but
it's not really useful for us. Their implementation acts as a UDP proxy
between the app and the statsd instance; they use the wildcard to signal
that a whole class of metrics should be dropped, or should be kept. We
don't need to do that, so it's not relevant to us.