Optimize the conversion of rules into regexes
The "split on period" hack was copied from seatgeek's implementation.

Prior to PR #1, the compiler assumed every string between `.`s was
either exactly one literal string, or exactly one placeholder.

That PR changed things so that you could put a literal string next to a
placeholder, without separating them with a period.

The regex approach in that PR kept the "split on `.`" functionality for
simplicity's sake, but it's not necessary - we can find all the
placeholders in one invocation of the regex matcher.